Task #4887: Merged trunk into the branch

+++ branches/Cobalt-Task4887/RTCP/Cobalt/InputProc/src/Station/RSP.h
@@ -100,7 +100,7 @@ namespace LOFAR
       union Payload {
         char data[8130];
 
-        // samples are structured as samples[nrBlocks][nrBeamlets],
+        // samples are structured as samples[nrBeamlets][nrBlocks],
         // so first all blocks of the first beamlet, then all blocks of the second
         // beamlet, etc.
         //
@@ -240,8 +240,10 @@ namespace LOFAR
       // decode the 4-bit complex type.
       static std::complex<int> decode4bit( int8 sample )
       {
-        int8 re = (sample << 4) >> 4; // preserve sign
-        int8 im = (sample     ) >> 4; // preserve sign
+        // intermediate after << will be int, not int8,
+        // so cast to get a signed int8 value.
+        int8 re = (int8)(sample << 4) >> 4; // preserve sign
+        int8 im =       (sample     ) >> 4; // preserve sign
 
         // balance range to [-7..7], subject to change!
         if (re == -8) re = -7;
